Fix read/write calls in xfs_io to allow buffers larger than 4GiB on 64 bit platforms. Merge of master-melb:xfs-cmds:24039a by kenmcd.

+++ include/input.h
@@ -41,8 +41,7 @@ extern char	**breakline(char *input, int *count);
 extern void	doneline(char *input, char **vec);
 extern char	*fetchline(void);
 
-extern long long cvtnum(unsigned int blocksize,
-			unsigned int sectorsize, char *s);
+extern long long cvtnum(size_t blocksize, size_t sectorsize, char *s);
 extern void	cvtstr(double value, char *str, size_t sz);
 extern unsigned long cvttime(char *s);
